Arm Improvements and Interface Change
These changes include replacing the big plus that covers the whole
screen with a small crosshair. It also primarily involves changes to the
arm, which I've created a new version of. I also created a method in the
Vertex class that allows AdjustableEdges to be retrieved without also
retrieving the regular Edges.